Refactor display backend: Extract hardware management into separate module
- Moved GPU memory introspection, Wayland output management, and display power control from `dispmanx_backend.py` to a new `hardware.py` module.
- Introduced `GpuInfo`, `WlrOutput`, and `DisplayPower` classes for better encapsulation and testability.
- Updated `Pi3dBackend` to utilize the new hardware management classes, simplifying the code and improving maintainability.
- Adjusted tests to reflect the new structure and ensure proper mocking of the hardware interactions.
