+++ src/metixel/backend/state.py
@@ -15,11 +15,11 @@ import threading
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
+from metixel.backend.system_metrics import SystemMetrics
 from metixel.shared.config import Config
 from metixel.shared.io import atomic_write_json
 from metixel.shared.models import MediaItem
 from metixel.shared.paths import install_root
-from metixel.shared.system_stats import read_meminfo
 
 if TYPE_CHECKING:
     from metixel.backend.processing.journal import ProcessingJournal
@@ -49,8 +49,15 @@ class StateManager:
         # "Clear cache" wipes it too.
         self._journal: ProcessingJournal | None = None
 
+        # System metrics service — owns all CPU/memory/disk/cache sizing.
+        self._metrics = SystemMetrics(config_provider=self._provide_config)
+
         # Ensure run directory exists
         run_dir.mkdir(parents=True, exist_ok=True)
+
+    def _provide_config(self) -> Config:
+        """Config provider for the metrics service (always the latest copy)."""
+        return self.config
 
     # -- Processing journal -------------------------------------------------
 
@@ -153,249 +160,12 @@ class StateManager:
     # -- System State --------------------------------------------------------
 
     def get_system_health(self) -> dict[str, Any]:
-        """Return system health metrics for MQTT/API reporting."""
-        import shutil
+        """Return system health metrics for MQTT/API reporting.
 
-        disk = shutil.disk_usage("/")
-        # On Linux, disk.free is space available to non-root users
-        # (excludes reserved blocks).  Compute "used" as total − free
-        # so the dashboard math is internally consistent:
-        #   free + used = total,  used / total = percent
-        used_bytes = disk.total - disk.free
-        cache_size = self._get_cache_size()
-        # Media size: ALL files under media/ (not just watched folders).
-        # Playlist counts: only items in enabled watch folders.
-        media_dir = Path(self.config.system.get("media_dir", "media/"))
-        if not media_dir.is_absolute():
-            media_dir = Path("/opt/metixel") / media_dir
-        media_size = self._get_media_folder_size(media_dir)
-        img_count, vid_count = self._get_playlist_counts()
-
-        cpu_pct = self._get_cpu_percent()
-        cpu_temp = self._get_cpu_temp()
-        mem = self._get_memory_stats()
-        swap = self._get_swap_stats()
-
-        return {
-            "uptime_seconds": self._get_uptime(),
-            "disk_total_gb": round(disk.total / (1024**3), 1),
-            "disk_used_gb": round(used_bytes / (1024**3), 1),
-            "disk_free_gb": round(disk.free / (1024**3), 1),
-            "disk_used_percent": round(used_bytes / disk.total * 100, 1),
-            "cache_size_mb": round(cache_size / (1024**2), 1),
-            "cache_size_bytes": cache_size,
-            "media_size_bytes": media_size,
-            "playlist_image_count": img_count,
-            "playlist_video_count": vid_count,
-            "cpu_percent": cpu_pct,
-            "cpu_temp_c": cpu_temp,
-            "memory_percent": mem["percent"],
-            "memory_used_gb": mem["used_gb"],
-            "memory_total_gb": mem["total_gb"],
-            "swap_percent": swap["percent"],
-            "swap_used_gb": swap["used_gb"],
-            "swap_total_gb": swap["total_gb"],
-        }
-
-    def _get_cache_size(self) -> int:
-        """Calculate the total size of the cache directory in bytes.
-
-        Returns 0 if the directory does not exist or cannot be read.
+        Delegates to the :class:`~metixel.backend.system_metrics.SystemMetrics`
+        service, which owns all CPU/memory/swap/disk/cache sizing.
         """
-        try:
-            config = self.config
-            cache_dir = Path(config.system.get("cache_dir", "cache/"))
-            if not cache_dir.is_absolute():
-                cache_dir = Path("/opt/metixel") / cache_dir
-            if not cache_dir.is_dir():
-                return 0
-            total = 0
-            for entry in cache_dir.rglob("*"):
-                if entry.is_file():
-                    with contextlib.suppress(OSError):
-                        total += entry.stat().st_size
-            return total
-        except Exception:
-            logger.debug("Could not compute cache size", exc_info=True)
-            return 0
-
-    def _get_media_folder_size(self, media_dir: Path) -> int:
-        """Calculate the total size of ALL files under *media_dir* in bytes.
-
-        This includes everything — watched folders, Immich sync data,
-        sample media — not just the actively-watched paths.
-        """
-        try:
-            if not media_dir.is_dir():
-                return 0
-            total = 0
-            seen: set[int] = set()
-            for entry in media_dir.rglob("*"):
-                if not entry.is_file():
-                    continue
-                try:
-                    st = entry.stat()
-                    if st.st_ino not in seen:
-                        seen.add(st.st_ino)
-                        total += st.st_size
-                except OSError:
-                    pass
-            return total
-        except Exception:
-            logger.debug("Could not compute media folder size", exc_info=True)
-            return 0
-
-    def _get_playlist_counts(self) -> tuple[int, int]:
-        """Count images and videos in enabled watch folders (playlist items).
-
-        This counts only items picked up by the folder watcher — not all
-        media on disk (Immich sync files, sample media, etc.).
-
-        Returns:
-            A tuple of ``(image_count, video_count)``.
-        """
-        IMG_EXT = {".jpg", ".jpeg", ".png", ".bmp", ".gif", ".webp"}  # noqa: N806
-        VID_EXT = {".mp4", ".mov", ".avi", ".mkv", ".webm", ".m4v", ".mpg", ".mpeg"}  # noqa: N806
-        try:
-            from metixel.shared.config import resolve_watch_paths
-
-            watch_paths = resolve_watch_paths(self.config)
-            img_count = 0
-            vid_count = 0
-            for media_folder in watch_paths:
-                if not media_folder.is_dir():
-                    continue
-                for entry in media_folder.rglob("*"):
-                    if not entry.is_file():
-                        continue
-                    suffix = entry.suffix.lower()
-                    if suffix in IMG_EXT:
-                        img_count += 1
-                    elif suffix in VID_EXT:
-                        vid_count += 1
-            return img_count, vid_count
-        except Exception:
-            logger.debug("Could not count media files", exc_info=True)
-            return 0, 0
-
-    @staticmethod
-    def _get_uptime() -> float:
-        """Get system uptime in seconds."""
-        try:
-            with open("/proc/uptime") as f:
-                return float(f.readline().split()[0])
-        except Exception:
-            return 0.0
-
-    # Cached /proc/stat from previous call for CPU delta calculation.
-    _prev_cpu_jiffies: float | None = None
-    _prev_idle_jiffies: float | None = None
-
-    @classmethod
-    def _get_cpu_percent(cls) -> float:
-        """Compute CPU utilisation as a percentage (0–100).
-
-        Reads ``/proc/stat`` and computes the delta in total CPU jiffies
-        since the previous call.  Because the web dashboard polls health
-        every ~3 seconds, the delta window matches the display interval.
-        On the first call (no prior sample) returns 0.0.
-        """
-        try:
-            with open("/proc/stat") as f:
-                line = f.readline()
-            # cpu  user nice system idle iowait irq softirq steal ...
-            parts = line.split()
-            if parts[0] != "cpu":
-                return 0.0
-            # Sum all jiffy columns (ignore guest/guest_nice)
-            jiffies = sum(int(x) for x in parts[1:8])
-            idle_now = int(parts[4])  # idle column
-        except Exception:
-            return 0.0
-
-        prev_total = cls._prev_cpu_jiffies
-        prev_idle = cls._prev_idle_jiffies
-        cls._prev_cpu_jiffies = float(jiffies)
-        cls._prev_idle_jiffies = float(idle_now)
-
-        if prev_total is None or prev_idle is None:
-            return 0.0
-
-        delta_total = jiffies - prev_total
-        if delta_total <= 0:
-            return 0.0
-
-        delta_idle = idle_now - prev_idle
-        if delta_idle < 0:
-            delta_idle = 0.0
-
-        pct = (1.0 - delta_idle / delta_total) * 100.0
-        return round(max(0.0, min(100.0, pct)), 1)
-
-    @staticmethod
-    def _get_cpu_temp() -> float:
-        """Read CPU temperature via ``vcgencmd measure_temp``.
-
-        Returns temperature in degrees Celsius, or 0.0 on failure.
-        """
-        try:
-            import subprocess
-
-            result = subprocess.run(
-                ["vcgencmd", "measure_temp"],
-                capture_output=True,
-                text=True,
-                timeout=5,
-            )
-            # Output format: "temp=45.0'C\n"
-            out = result.stdout.strip()
-            if out.startswith("temp="):
-                return float(out.split("=")[1].split("'")[0])
-        except Exception:
-            pass
-        return 0.0
-
-    @staticmethod
-    def _get_memory_stats() -> dict[str, float]:
-        """Read ``/proc/meminfo`` and return memory usage stats.
-
-        Returns:
-            Dict with keys ``percent``, ``used_gb``, ``total_gb``.
-            Falls back to zeros if ``/proc/meminfo`` cannot be read.
-        """
-        mem = read_meminfo()
-        total_kb = mem.get("MemTotal", 0)
-        available_kb = mem.get("MemAvailable", 0)
-        if total_kb <= 0:
-            return {"percent": 0.0, "used_gb": 0.0, "total_gb": 0.0}
-        used_kb = total_kb - available_kb
-        pct = round(used_kb / total_kb * 100.0, 1)
-        return {
-            "percent": max(0.0, min(100.0, pct)),
-            "used_gb": round(used_kb / (1024 * 1024), 1),
-            "total_gb": round(total_kb / (1024 * 1024), 1),
-        }
-
-    @staticmethod
-    def _get_swap_stats() -> dict[str, float]:
-        """Read ``/proc/meminfo`` and return swap usage stats.
-
-        Returns:
-            Dict with keys ``percent``, ``used_gb``, ``total_gb``.
-            Falls back to zeros if swap is disabled or unreadable.
-        """
-        mem = read_meminfo()
-        total_kb = mem.get("SwapTotal", 0)
-        free_kb = mem.get("SwapFree", 0)
-        if total_kb <= 0:
-            return {"percent": 0.0, "used_gb": 0.0, "total_gb": 0.0}
-        used_kb = total_kb - free_kb
-        pct = round(used_kb / total_kb * 100.0, 1)
-        return {
-            "percent": max(0.0, min(100.0, pct)),
-            "used_gb": round(used_kb / (1024 * 1024), 1),
-            "total_gb": round(total_kb / (1024 * 1024), 1),
-        }
+        return self._metrics.get_system_health()
 
     # -- Playlist Management -------------------------------------------------
